pyproject.toml
[tool.poetry]
name = "my-wiki"
version = "0.1.0"
description = ""
authors = ["Creeper Eye <iostream1204@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.13"
django = "^5.1"
psycopg2 = "^2.9.9"
uwsgi = "^2.0.26"
djangorestframework = "^3.15.2"
django-webpack-loader = "^3.1.0"
psutil = "^6.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
